organização de estilo do Search
